Remove Dart sync client
