+++ packages/powersync/test/sync/stream_test.dart
@@ -274,23 +274,4 @@ void main() {
       ),
     );
   });
-
-  test('passes app metadata to the server (legacy sync client)', () async {
-    options = SyncOptions(
-      // ignore: deprecated_member_use_from_same_package
-      syncImplementation: SyncClientImplementation.dart,
-      appMetadata: {'foo': 'bar'},
-    );
-
-    await waitForConnection();
-
-    final request = await syncService.waitForListener;
-    expect(
-      json.decode(await request.readAsString()),
-      containsPair(
-        'app_metadata',
-        containsPair('foo', 'bar'),
-      ),
-    );
-  });
 }
